requirements.txt
Django==3.1
factory-boy==3.2.0
Faker==5.6.0
h11==0.8.1
h2==3.2.0
http3==0.6.7
idna==2.8
Paperboy==1.0.1
restless==2.1.1
text-unidecode==1.3

